Gui
The gui is now working and the next step would now be too find a way to limit the error of small little blobs that seem too appear and random.

As well the ability to network between the kangaroo and the roborio should be explored in order to make sure that everything will work in connection.

As well the main function should be finished for the reflective tape will have to be finished.